use uintptr_t for TouchId

+++ Include/RmlUi/Core/Types.h
@@ -101,7 +101,7 @@ enum class EventId : uint16_t;
 enum class PropertyId : uint8_t;
 enum class MediaQueryId : uint8_t;
 enum class FamilyId : int;
-using TouchId = int;
+using TouchId = uintptr_t;
 
 // Types for external interfaces.
 using FileHandle = uintptr_t;
